Balance/src/requirements.txt
importlib-metadata==6.6.0
importlib-resources==5.12.0
libusb==1.0.26b5
packaging==23.1
Pillow==9.5.0
pkg-about==1.0.8
pyserial==3.5
python-barcode==0.14.0
pyusb==1.2.1
tk==0.1.0
tomli==2.0.1
zipp==3.15.0